Tested and works
Used repl.it to confirm it works properly. This doubly-linked list will read a name from an input.txt delete it if delete precedes the name, display that it was detected and that it will be deleted. And then display in both ascending and descending order, saving the results to a file named output.txt